Add front-end interface to backend

+++ server/src/config/reading.rs
@@ -2,9 +2,9 @@ use serde::{Serialize, Deserialize};
 
 use super::ActionConfig;
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Clone)]
 pub struct ReadingConfig {
     pub title: String,
-    pub period_ms: usize,
+    pub period_ms: u64,
     pub on_period: ActionConfig,
 }
